requirements.txt
pyserial==3.5
pytest==8.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
